requirements.txt
snowflake-connector-python==3.5.0
python-dotenv==1.0.0 